+++ R/zzz.R
@@ -1,6 +1,5 @@
 # nocov start
 .onLoad <- function(...) {
-  options(warnPartialMatchDollar = TRUE)
   Sys.setenv("_R_CHECK_LENGTH_1_CONDITION_" = "true")
 
   .setEsqlabsColors()